+++ ast/disable_tail_calls_attr.go
@@ -1,5 +1,7 @@
 package ast
 
+// DisableTailCallAttr is a type of attribute that is optionally attached to a variable
+// or struct field definition.
 type DisableTailCallsAttr struct {
 	Addr       Address
 	Pos        Position
