+++ ast/util.go
@@ -1,6 +1,9 @@
 package ast
 
 import (
+	"bytes"
+	"encoding/json"
+	"fmt"
 	"strconv"
 	"strings"
 )
@@ -32,4 +35,42 @@ func atof(s string) float64 {
 	}
 
 	return f
+}
+
+// Atos - ASTree to string
+// Typically using for debug
+func Atos(node Node) string {
+	j, err := json.Marshal(node)
+	if err != nil {
+		panic(err)
+	}
+	var out bytes.Buffer
+	err = json.Indent(&out, j, "", "\t")
+	if err != nil {
+		panic(err)
+	}
+	var str string
+	str += fmt.Sprint("==== START OF AST tree ====\n")
+	str += out.String()
+	str += fmt.Sprintf("\nTypes tree:\n")
+	str += typesTree(node, 0)
+	str += fmt.Sprint("==== END OF AST tree ====\n")
+	return str
+}
+
+func typesTree(node Node, depth int) (str string) {
+	if node == (Node)(nil) {
+		return ""
+	}
+	for i := 0; i < depth; i++ {
+		str += "\t"
+	}
+	str += fmt.Sprintf("%T\n", node)
+	depth++
+	if len(node.Children()) > 0 {
+		for _, n := range node.Children() {
+			str += typesTree(n, depth)
+		}
+	}
+	return str
 }
